[Chore][Workloads] merge fragmented per-op workload files by family
16 single-class workload files (9-22 lines each) collapse into three
family modules, mirroring the manifest family layout:

- workloads/normalization.py <- rms_norm, layer_norm, fused_add_*,
  ada_layer_norm{,_zero}, group_norm, instance_norm, batch_norm
- workloads/reduction.py <- reduce, argreduce, softmax, vector_norm,
  logical_reduce
- workloads/elementwise.py <- activation, binary_arith

Class names, bodies, and behavior unchanged; consumer imports
retargeted. Collected test node set is byte-identical.
