+++ benchmarks/ops/bench_argreduce.py
@@ -12,7 +12,7 @@ import torch
 from benchmarks.benchmark_base import BenchmarkReport, ManifestBenchmark, workloads_to_params
 from tileops.ops.reduction.argmax import ArgmaxFwdOp
 from tileops.ops.reduction.argmin import ArgminFwdOp
-from workloads.argreduce import ArgmaxTest, ArgminTest
+from workloads.reduction import ArgmaxTest, ArgminTest
 
 _ARGMAX_OP = "ArgmaxFwdOp"
 _ARGMIN_OP = "ArgminFwdOp"
