+++ benchmarks/ops/bench_fused_add_rms_norm.py
@@ -4,7 +4,7 @@ import torch
 from benchmarks.benchmark_base import BenchmarkReport, ManifestBenchmark
 from tileops.manifest import load_workloads
 from tileops.ops.norm.fused_add_rms_norm import FusedAddRMSNormFwdOp
-from workloads.fused_add_rms_norm import FusedAddRMSNormTest
+from workloads.normalization import FusedAddRMSNormTest
 
 _OP_NAME = "FusedAddRMSNormFwdOp"
