+++ tests/ops/test_activation.py
@@ -9,7 +9,7 @@ import torch.nn.functional as F
 
 from tests.test_base import FixtureBase, TestBase
 from tileops.ops.elementwise import ReluFwdOp
-from workloads.activation import ReluTest as _ReluTestWorkload
+from workloads.elementwise import ReluTest as _ReluTestWorkload
 
 
 class ReluTest(_ReluTestWorkload, TestBase):
