+++ tests/ops/test_fused_add_layer_norm.py
@@ -4,7 +4,7 @@ import torch.nn.functional as F
 
 from tests.test_base import FixtureBase, TestBase
 from tileops.ops.norm.fused_add_layer_norm import FusedAddLayerNormFwdOp
-from workloads.fused_add_layer_norm import (
+from workloads.normalization import (
     FusedAddLayerNormTest as _FusedAddLayerNormTestWorkload,
 )
