--- workloads/activation.py
@@ -1,14 +1,0 @@
-import torch
-
-from workloads.workload_base import WorkloadBase
-
-
-class ReluTest(WorkloadBase):
-
-    def __init__(self, n_total: int, dtype: torch.dtype):
-        self.n_total = n_total
-        self.dtype = dtype
-
-    def gen_inputs(self) -> tuple[torch.Tensor]:
-        x = torch.randn(self.n_total, dtype=self.dtype, device="cuda")
-        return (x,)
