--- workloads/argreduce.py
@@ -1,9 +1,0 @@
-from workloads.workload_base import RandnTest
-
-
-class ArgmaxTest(RandnTest):
-    """Workload definition for ArgmaxFwdOp."""
-
-
-class ArgminTest(RandnTest):
-    """Workload definition for ArgminFwdOp."""
