--- workloads/rms_norm.py
@@ -1,17 +1,0 @@
-import torch
-
-from workloads.workload_base import WorkloadBase
-
-
-class RMSNormTest(WorkloadBase):
-
-    def __init__(self, m: int, n: int, dtype: torch.dtype, eps: float = 1e-6):
-        self.m = m
-        self.n = n
-        self.dtype = dtype
-        self.eps = eps
-
-    def gen_inputs(self) -> tuple[torch.Tensor, torch.Tensor]:
-        x = torch.randn(self.m, self.n, dtype=self.dtype, device="cuda")
-        weight = torch.randn(self.n, dtype=self.dtype, device="cuda")
-        return x, weight
