--- workloads/vector_norm.py
@@ -1,13 +1,0 @@
-from workloads.workload_base import RandnTest
-
-
-class L1NormTest(RandnTest):
-    """Workload definition for L1NormFwdOp."""
-
-
-class L2NormTest(RandnTest):
-    """Workload definition for L2NormFwdOp."""
-
-
-class InfNormTest(RandnTest):
-    """Workload definition for InfNormFwdOp."""
